New demo from John: CallbackMaker.
This demo shows the use of every callback that you can register with
freeglut, and also generates event reports so that you can see what is
happening to the program as it runs.

Not much to look at, but both utilitarian and a practical example.

Please double-check that I updated everything that needs to be updated.
I reran autogen.sh and ./configure, and it built okay for me.  (^&


git-svn-id: http://svn.code.sf.net/p/freeglut/code/trunk/freeglut/freeglut@332 7f0cb862-5218-0410-a997-914c9d46530a